+++ Plugins/FreeRDP/Client/PluginFreeRDP.h
@@ -32,7 +32,7 @@ public:
     virtual const QString Version() const override;
     
 private:
-    virtual CConnecter* CreateConnecter(const QString& szId) override;
+    virtual CConnecter* OnCreateConnecter(const QString& szId) override;
 };
 
 #endif // CPluginFreeRdp_H_KL_2021_07_23